test: improve router test suite performance
- Make most tests parallel
- Avoid generating the same configuration multiple times
- Make timings more tighter when possible, to make the tests not spend
so much time waiting

This reduces the runtime for the whole testsuite from ~180s to ~18s
