+++ router-tests/websocket_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/wundergraph/cosmo/router/core"
 	"net"
 	"net/http"
 	"testing"
@@ -15,6 +14,7 @@ import (
 	"github.com/hasura/go-graphql-client/pkg/jsonutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/wundergraph/cosmo/router/core"
 )
 
 type wsMessage struct {
@@ -68,99 +68,213 @@ func connectedWebsocket(tb testing.TB, serverPort int, serverPath string, opts *
 	return conn
 }
 
-func TestQueryOverWebsocket(t *testing.T) {
-	const (
-		query           = `{ employees { id } }`
-		expectedPayload = `{"data":{"employees":[{"id":1},{"id":2},{"id":3},{"id":4},{"id":5},{"id":7},{"id":8},{"id":9},{"id":10},{"id":11},{"id":12}]}}`
-	)
-	_, port := setupListeningServer(t)
-	conn := connectedWebsocket(t, port, "/graphql", nil)
-	var err error
-	q := &testQuery{
-		Body: query,
-	}
-	const messageID = "1"
-	err = conn.WriteJSON(&wsMessage{
-		ID:      messageID,
-		Type:    "subscribe",
-		Payload: q.Data(),
-	})
-	assert.NoError(t, err)
-	var msg wsMessage
-	err = connReadJSON(conn, &msg)
-	require.NoError(t, err)
-	assert.Equal(t, "next", msg.Type)
-	assert.Equal(t, messageID, msg.ID)
-	// Delete any "extensions" field from the payload, we don't care about it for now
-	var payload map[string]json.RawMessage
-	err = json.Unmarshal(msg.Payload, &payload)
-	require.NoError(t, err)
-	delete(payload, "extensions")
-	payloadBytes, err := json.Marshal(payload)
-	require.NoError(t, err)
-	assert.Equal(t, expectedPayload, string(payloadBytes))
-	err = connReadJSON(conn, &msg)
-	require.NoError(t, err)
-	assert.Equal(t, "complete", msg.Type)
-	assert.Equal(t, messageID, msg.ID)
-}
+func TestWebsockets(t *testing.T) {
+	t.Parallel()
 
-func TestSubscriptionOverWebsocket(t *testing.T) {
-	type currentTimePayload struct {
-		Data struct {
+	_, port := setupListeningServer(t)
+
+	t.Run("query", func(t *testing.T) {
+		const (
+			query           = `{ employees { id } }`
+			expectedPayload = `{"data":{"employees":[{"id":1},{"id":2},{"id":3},{"id":4},{"id":5},{"id":7},{"id":8},{"id":9},{"id":10},{"id":11},{"id":12}]}}`
+		)
+
+		conn := connectedWebsocket(t, port, "/graphql", nil)
+		var err error
+		q := &testQuery{
+			Body: query,
+		}
+		const messageID = "1"
+		err = conn.WriteJSON(&wsMessage{
+			ID:      messageID,
+			Type:    "subscribe",
+			Payload: q.Data(),
+		})
+		assert.NoError(t, err)
+		var msg wsMessage
+		err = connReadJSON(conn, &msg)
+		require.NoError(t, err)
+		assert.Equal(t, "next", msg.Type)
+		assert.Equal(t, messageID, msg.ID)
+		// Delete any "extensions" field from the payload, we don't care about it for now
+		var payload map[string]json.RawMessage
+		err = json.Unmarshal(msg.Payload, &payload)
+		require.NoError(t, err)
+		delete(payload, "extensions")
+		payloadBytes, err := json.Marshal(payload)
+		require.NoError(t, err)
+		assert.Equal(t, expectedPayload, string(payloadBytes))
+		err = connReadJSON(conn, &msg)
+		require.NoError(t, err)
+		assert.Equal(t, "complete", msg.Type)
+		assert.Equal(t, messageID, msg.ID)
+	})
+
+	t.Run("subscription", func(t *testing.T) {
+
+		type currentTimePayload struct {
+			Data struct {
+				CurrentTime struct {
+					UnixTime  float64 `json:"unixTime"`
+					Timestamp string  `json:"timestamp"`
+				} `json:"currentTime"`
+			} `json:"data"`
+		}
+
+		conn := connectedWebsocket(t, port, "/graphql", nil)
+		var err error
+		const messageID = "1"
+		err = conn.WriteJSON(&wsMessage{
+			ID:      messageID,
+			Type:    "subscribe",
+			Payload: []byte(`{"query":"subscription { currentTime { unixTime timeStamp }}"}`),
+		})
+		require.NoError(t, err)
+		var msg wsMessage
+		var payload currentTimePayload
+
+		// Read a result and store its timestamp, next result should be 1 second later
+		err = connReadJSON(conn, &msg)
+		require.NoError(t, err)
+		assert.Equal(t, messageID, msg.ID)
+		assert.Equal(t, "next", msg.Type)
+		err = json.Unmarshal(msg.Payload, &payload)
+		require.NoError(t, err)
+
+		unix1 := payload.Data.CurrentTime.UnixTime
+
+		err = connReadJSON(conn, &msg)
+		require.NoError(t, err)
+		assert.Equal(t, messageID, msg.ID)
+		assert.Equal(t, "next", msg.Type)
+		err = json.Unmarshal(msg.Payload, &payload)
+		require.NoError(t, err)
+
+		unix2 := payload.Data.CurrentTime.UnixTime
+		assert.Equal(t, unix1+1, unix2)
+
+		// Sending a complete must stop the subscription
+		err = conn.WriteJSON(&wsMessage{
+			ID:   messageID,
+			Type: "complete",
+		})
+		require.NoError(t, err)
+		conn.SetReadDeadline(time.Now().Add(2 * time.Second))
+		// This should timeout because no more data is coming
+		_, _, err = conn.NextReader()
+		netErr, ok := err.(net.Error)
+		require.True(t, ok, "error is not a net.Error, got %T = %v", err, err)
+		assert.True(t, netErr.Timeout())
+		conn.SetReadDeadline(time.Time{})
+	})
+
+	t.Run("error", func(t *testing.T) {
+
+		conn := connectedWebsocket(t, port, "/graphql", nil)
+		var err error
+		const messageID = "1"
+		err = conn.WriteJSON(&wsMessage{
+			ID:      messageID,
+			Type:    "subscribe",
+			Payload: []byte(`{"query":"subscription { does_not_exist }"}`),
+		})
+		require.NoError(t, err)
+		var msg wsMessage
+		err = connReadJSON(conn, &msg)
+		require.NoError(t, err)
+		assert.Equal(t, "error", msg.Type)
+		// Payload should be an array of GraphQLError
+		var errors []graphqlError
+		err = json.Unmarshal(msg.Payload, &errors)
+		require.NoError(t, err)
+		assert.Len(t, errors, 1)
+		assert.Equal(t, errors[0].Message, `field: does_not_exist not defined on type: Subscription`)
+	})
+
+	t.Run("subscription with library", func(t *testing.T) {
+		var subscription struct {
 			CurrentTime struct {
-				UnixTime  float64 `json:"unixTime"`
-				Timestamp string  `json:"timestamp"`
-			} `json:"currentTime"`
-		} `json:"data"`
-	}
-
-	_, port := setupListeningServer(t)
-	conn := connectedWebsocket(t, port, "/graphql", nil)
-	var err error
-	const messageID = "1"
-	err = conn.WriteJSON(&wsMessage{
-		ID:      messageID,
-		Type:    "subscribe",
-		Payload: []byte(`{"query":"subscription { currentTime { unixTime timeStamp }}"}`),
+				UnixTime  float64 `graphql:"unixTime"`
+				Timestamp string  `graphql:"timeStamp"`
+			} `graphql:"currentTime"`
+		}
+		subscriptionURL := fmt.Sprintf("ws://localhost:%d/graphql", port)
+		protocols := []graphql.SubscriptionProtocolType{
+			graphql.GraphQLWS,
+			graphql.SubscriptionsTransportWS,
+		}
+		for _, p := range protocols {
+			p := p
+			t.Run(string(p), func(t *testing.T) {
+				client := graphql.NewSubscriptionClient(subscriptionURL).WithProtocol(p)
+				t.Cleanup(func() {
+					err := client.Close()
+					assert.NoError(t, err)
+				})
+				var firstTime float64
+				subscriptionID, err := client.Subscribe(&subscription, nil, func(dataValue []byte, errValue error) error {
+					require.NoError(t, errValue)
+					data := subscription
+					err := jsonutil.UnmarshalGraphQL(dataValue, &data)
+					require.NoError(t, err)
+					if firstTime == 0 {
+						firstTime = data.CurrentTime.UnixTime
+					} else {
+						assert.Equal(t, firstTime+1, data.CurrentTime.UnixTime)
+						return graphql.ErrSubscriptionStopped
+					}
+					return nil
+				})
+				require.NoError(t, err)
+				require.NotEqual(t, "", subscriptionID)
+				require.NoError(t, client.Run())
+			})
+		}
 	})
-	require.NoError(t, err)
-	var msg wsMessage
-	var payload currentTimePayload
 
-	// Read a result and store its timestamp, next result should be 1 second later
-	err = connReadJSON(conn, &msg)
-	require.NoError(t, err)
-	assert.Equal(t, messageID, msg.ID)
-	assert.Equal(t, "next", msg.Type)
-	err = json.Unmarshal(msg.Payload, &payload)
-	require.NoError(t, err)
+	t.Run("forward extensions", func(t *testing.T) {
 
-	unix1 := payload.Data.CurrentTime.UnixTime
+		// Make sure sending two simultaneous subscriptions with different extensions
+		// triggers two subscriptions to the upstream
+		conn1 := connectedWebsocket(t, port, "/graphql", nil)
+		conn2 := connectedWebsocket(t, port, "/graphql", nil)
+		var err error
+		err = conn1.WriteJSON(&wsMessage{
+			ID:      "1",
+			Type:    "subscribe",
+			Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }","extensions":{"token":"123"}}`),
+		})
+		require.NoError(t, err)
 
-	err = connReadJSON(conn, &msg)
-	require.NoError(t, err)
-	assert.Equal(t, messageID, msg.ID)
-	assert.Equal(t, "next", msg.Type)
-	err = json.Unmarshal(msg.Payload, &payload)
-	require.NoError(t, err)
+		err = conn2.WriteJSON(&wsMessage{
+			ID:      "2",
+			Type:    "subscribe",
+			Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }","extensions":{"token":"456"}}`),
+		})
+		require.NoError(t, err)
 
-	unix2 := payload.Data.CurrentTime.UnixTime
-	assert.Equal(t, unix1+1, unix2)
+		var msg wsMessage
+		var payload struct {
+			Data struct {
+				InitialPayload struct {
+					Extensions struct {
+						Token string `json:"token"`
+					} `json:"extensions"`
+				} `json:"initialPayload"`
+			} `json:"data"`
+		}
+		err = connReadJSON(conn1, &msg)
+		require.NoError(t, err)
+		err = json.Unmarshal(msg.Payload, &payload)
+		require.NoError(t, err)
+		assert.Equal(t, "123", payload.Data.InitialPayload.Extensions.Token)
 
-	// Sending a complete must stop the subscription
-	err = conn.WriteJSON(&wsMessage{
-		ID:   messageID,
-		Type: "complete",
+		err = connReadJSON(conn2, &msg)
+		require.NoError(t, err)
+		err = json.Unmarshal(msg.Payload, &payload)
+		require.NoError(t, err)
+		assert.Equal(t, "456", payload.Data.InitialPayload.Extensions.Token)
 	})
-	require.NoError(t, err)
-	conn.SetReadDeadline(time.Now().Add(2 * time.Second))
-	// This should timeout because no more data is coming
-	_, _, err = conn.NextReader()
-	netErr, ok := err.(net.Error)
-	require.True(t, ok, "error is not a net.Error, got %T = %v", err, err)
-	assert.True(t, netErr.Timeout())
-	conn.SetReadDeadline(time.Time{})
 }
 
 type graphqlErrorResponse struct {
@@ -171,30 +285,46 @@ type graphqlError struct {
 	Message string `json:"message"`
 }
 
-func TestErrorOverWebsocket(t *testing.T) {
-	_, port := setupListeningServer(t)
-	conn := connectedWebsocket(t, port, "/graphql", nil)
+func TestWsConnectionWithSameGraphQLPathAsPlayground(t *testing.T) {
+	t.Parallel()
+	_, port := setupListeningServer(t, core.WithGraphQLPath("/"))
+	conn := connectedWebsocket(t, port, "/", &connectedWebsocketOptions{
+		InitialPayload: map[string]any{"123": 456, "extensions": map[string]any{"hello": "world"}},
+	})
 	var err error
-	const messageID = "1"
 	err = conn.WriteJSON(&wsMessage{
-		ID:      messageID,
+		ID:      "1",
 		Type:    "subscribe",
-		Payload: []byte(`{"query":"subscription { does_not_exist }"}`),
+		Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }"}`),
 	})
 	require.NoError(t, err)
 	var msg wsMessage
 	err = connReadJSON(conn, &msg)
 	require.NoError(t, err)
-	assert.Equal(t, "error", msg.Type)
-	// Payload should be an array of GraphQLError
-	var errors []graphqlError
-	err = json.Unmarshal(msg.Payload, &errors)
+	assert.JSONEq(t, `{"data":{"initialPayload":{"123":456,"extensions":{"hello":"world"}}}}`, string(msg.Payload))
+}
+
+func TestWsConnectionWithDifferentGraphQLPath(t *testing.T) {
+	t.Parallel()
+	_, port := setupListeningServer(t, core.WithGraphQLPath("/foo"))
+	conn := connectedWebsocket(t, port, "/foo", &connectedWebsocketOptions{
+		InitialPayload: map[string]any{"123": 456, "extensions": map[string]any{"hello": "world"}},
+	})
+	var err error
+	err = conn.WriteJSON(&wsMessage{
+		ID:      "1",
+		Type:    "subscribe",
+		Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }"}`),
+	})
 	require.NoError(t, err)
-	assert.Len(t, errors, 1)
-	assert.Equal(t, errors[0].Message, `field: does_not_exist not defined on type: Subscription`)
+	var msg wsMessage
+	err = connReadJSON(conn, &msg)
+	require.NoError(t, err)
+	assert.JSONEq(t, `{"data":{"initialPayload":{"123":456,"extensions":{"hello":"world"}}}}`, string(msg.Payload))
 }
 
 func TestShutdownWithActiveWebsocket(t *testing.T) {
+	t.Parallel()
 	server, port := setupListeningServer(t)
 	conn := connectedWebsocket(t, port, "/graphql", nil)
 	var err error
@@ -219,129 +349,8 @@ func TestShutdownWithActiveWebsocket(t *testing.T) {
 	assert.Equal(t, websocket.CloseAbnormalClosure, cerr.Code)
 }
 
-func TestSubscriptionsOverWebsocketLibrary(t *testing.T) {
-	var subscription struct {
-		CurrentTime struct {
-			UnixTime  float64 `graphql:"unixTime"`
-			Timestamp string  `graphql:"timeStamp"`
-		} `graphql:"currentTime"`
-	}
-	_, port := setupListeningServer(t)
-	subscriptionURL := fmt.Sprintf("ws://localhost:%d/graphql", port)
-	protocols := []graphql.SubscriptionProtocolType{
-		graphql.GraphQLWS,
-		graphql.SubscriptionsTransportWS,
-	}
-	for _, p := range protocols {
-		p := p
-		t.Run(string(p), func(t *testing.T) {
-			client := graphql.NewSubscriptionClient(subscriptionURL).WithProtocol(p)
-			t.Cleanup(func() {
-				err := client.Close()
-				assert.NoError(t, err)
-			})
-			var firstTime float64
-			subscriptionID, err := client.Subscribe(&subscription, nil, func(dataValue []byte, errValue error) error {
-				require.NoError(t, errValue)
-				data := subscription
-				err := jsonutil.UnmarshalGraphQL(dataValue, &data)
-				require.NoError(t, err)
-				if firstTime == 0 {
-					firstTime = data.CurrentTime.UnixTime
-				} else {
-					assert.Equal(t, firstTime+1, data.CurrentTime.UnixTime)
-					return graphql.ErrSubscriptionStopped
-				}
-				return nil
-			})
-			require.NoError(t, err)
-			require.NotEqual(t, "", subscriptionID)
-			require.NoError(t, client.Run())
-		})
-	}
-}
-
-func TestExtensionsForwardingOverWebsocket(t *testing.T) {
-	// Make sure sending two simultaneous subscriptions with different extensions
-	// triggers two subscriptions to the upstream
-	_, port := setupListeningServer(t)
-	conn1 := connectedWebsocket(t, port, "/graphql", nil)
-	conn2 := connectedWebsocket(t, port, "/graphql", nil)
-	var err error
-	err = conn1.WriteJSON(&wsMessage{
-		ID:      "1",
-		Type:    "subscribe",
-		Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }","extensions":{"token":"123"}}`),
-	})
-	require.NoError(t, err)
-
-	err = conn2.WriteJSON(&wsMessage{
-		ID:      "2",
-		Type:    "subscribe",
-		Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }","extensions":{"token":"456"}}`),
-	})
-	require.NoError(t, err)
-
-	var msg wsMessage
-	var payload struct {
-		Data struct {
-			InitialPayload struct {
-				Extensions struct {
-					Token string `json:"token"`
-				} `json:"extensions"`
-			} `json:"initialPayload"`
-		} `json:"data"`
-	}
-	err = connReadJSON(conn1, &msg)
-	require.NoError(t, err)
-	err = json.Unmarshal(msg.Payload, &payload)
-	require.NoError(t, err)
-	assert.Equal(t, "123", payload.Data.InitialPayload.Extensions.Token)
-
-	err = connReadJSON(conn2, &msg)
-	require.NoError(t, err)
-	err = json.Unmarshal(msg.Payload, &payload)
-	require.NoError(t, err)
-	assert.Equal(t, "456", payload.Data.InitialPayload.Extensions.Token)
-}
-
-func TestWsConnectionWithSameGraphQLPathAsPlayground(t *testing.T) {
-	_, port := setupListeningServer(t, core.WithGraphQLPath("/"))
-	conn := connectedWebsocket(t, port, "/", &connectedWebsocketOptions{
-		InitialPayload: map[string]any{"123": 456, "extensions": map[string]any{"hello": "world"}},
-	})
-	var err error
-	err = conn.WriteJSON(&wsMessage{
-		ID:      "1",
-		Type:    "subscribe",
-		Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }"}`),
-	})
-	require.NoError(t, err)
-	var msg wsMessage
-	err = connReadJSON(conn, &msg)
-	require.NoError(t, err)
-	assert.JSONEq(t, `{"data":{"initialPayload":{"123":456,"extensions":{"hello":"world"}}}}`, string(msg.Payload))
-}
-
-func TestWsConnectionWithDifferentGraphQLPath(t *testing.T) {
-	_, port := setupListeningServer(t, core.WithGraphQLPath("/foo"))
-	conn := connectedWebsocket(t, port, "/foo", &connectedWebsocketOptions{
-		InitialPayload: map[string]any{"123": 456, "extensions": map[string]any{"hello": "world"}},
-	})
-	var err error
-	err = conn.WriteJSON(&wsMessage{
-		ID:      "1",
-		Type:    "subscribe",
-		Payload: []byte(`{"query":"subscription { initialPayload(repeat:3) }"}`),
-	})
-	require.NoError(t, err)
-	var msg wsMessage
-	err = connReadJSON(conn, &msg)
-	require.NoError(t, err)
-	assert.JSONEq(t, `{"data":{"initialPayload":{"123":456,"extensions":{"hello":"world"}}}}`, string(msg.Payload))
-}
-
 func TestExtensionsForwardingOverWebsocketWithInitialPayload(t *testing.T) {
+	t.Parallel()
 	_, port := setupListeningServer(t)
 	t.Run("single connection with initial payload", func(t *testing.T) {
 		conn := connectedWebsocket(t, port, "/graphql", &connectedWebsocketOptions{
